updateDelete in 1

+++ app/controllers/ServiceController.php
@@ -23,13 +23,27 @@ class ServiceController extends BaseController {
 		->with('message', 'A service has been renamed to ' . $service_name . '.');
 	}
 	
-	public function postDelete() {
-		$service_ids = Input::get('service_delete');
+	public function postUpdatedelete() {
+		$for_deletion = Input::get('service_delete');
 		$Service = new Service();
 		$delete_message = '';
-		foreach ($service_ids as $key => $service_id) {
-			$delete_message .= $Service->serviceName($service_id) . ' has been removed.';
-			$Service->deleteService($service_id);
+		$update_message = '';
+		$service_ids = $Service->fetchServices();
+		foreach ($service_ids as $service_id => $service_name) {
+			$count = 0;
+						
+			// update the status of the service (active/inactive)
+			$status = Input::get('service_status_' . $service_id);
+			$Service->updateStatus($status, $service_id);
+			$update_message .= 'The status of '. $service_name . ' has been updated.<br/>';
+			
+			while ($count < count($for_deletion)) {
+				if ($service_id == $for_deletion[$count]) {
+					$delete_message .= $service_name . ' has been removed.<br/>';
+					$Service->deleteService($service_id);
+				}
+				$count++;
+			}
 		}
 		return Redirect::to('user/dashboard')
 			->with('message', $delete_message);
